Tweaks styling of proj. "commits" list to increase space available.
Prior to these tweaks, the commit subject was cut off at 50 characters because that's all that was likely to fit. As of these tweaks, the subject is now cut off at the standard 72 characters.

This could definitely use more work, but the primary goal was to get the full subject to show, and this takes care of that.